added properties to evolution plots

+++ param_evolution.py
@@ -610,7 +610,7 @@ for reg in halos_in_pop:
                 haloids.append(grp)
                 nbh.append(len(bh_id_dict[snap][reg][bh_id_dict[snap][reg] == grp]))
 
-        if all(np.array(sfrs[np.array(zs) <= 5]) > 0.1):
+        if all(np.array(sfrs)[np.array(zs, dtype=float) <= 5] > 0.1):
             continue
 
         print("====================", reg, "====================")
